+++ src/main/java/ru/endlesscode/rpginventory/RPGInventory.java
@@ -25,13 +25,19 @@ import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.RegisteredServiceProvider;
+import org.bukkit.plugin.ServicePriority;
+import org.bukkit.plugin.ServicesManager;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import ru.endlesscode.inspector.bukkit.command.TrackedCommandExecutor;
 import ru.endlesscode.inspector.bukkit.plugin.PluginLifecycle;
 import ru.endlesscode.inspector.bukkit.scheduler.TrackedBukkitRunnable;
+import ru.endlesscode.mimic.classes.BukkitClassSystem;
+import ru.endlesscode.mimic.items.BukkitItemsRegistry;
+import ru.endlesscode.mimic.level.BukkitLevelSystem;
 import ru.endlesscode.rpginventory.compat.VersionHandler;
+import ru.endlesscode.rpginventory.compat.mimic.RPGInventoryItemsRegistry;
 import ru.endlesscode.rpginventory.compat.mypet.MyPetManager;
 import ru.endlesscode.rpginventory.event.listener.ArmorEquipListener;
 import ru.endlesscode.rpginventory.event.listener.ElytraListener;
@@ -53,7 +59,6 @@ import ru.endlesscode.rpginventory.pet.PetManager;
 import ru.endlesscode.rpginventory.resourcepack.ResourcePackModule;
 import ru.endlesscode.rpginventory.utils.Log;
 import ru.endlesscode.rpginventory.utils.PlayerUtils;
-import ru.endlesscode.rpginventory.utils.SafeEnums;
 import ru.endlesscode.rpginventory.utils.StringUtils;
 import ru.endlesscode.rpginventory.utils.Version;
 
@@ -65,8 +70,8 @@ public class RPGInventory extends PluginLifecycle {
     private Permission perms;
     private Economy economy;
 
-    private PlayerUtils.LevelSystem levelSystem;
-    private PlayerUtils.ClassSystem classSystem;
+    private BukkitLevelSystem.Provider levelSystemProvider;
+    private BukkitClassSystem.Provider classSystemProvider;
 
     private FileLanguage language;
     private boolean placeholderApiHooked = false;
@@ -104,12 +109,12 @@ public class RPGInventory extends PluginLifecycle {
         return instance.myPetHooked;
     }
 
-    public static PlayerUtils.LevelSystem getLevelSystem() {
-        return instance.levelSystem;
+    public static BukkitLevelSystem getLevelSystem(@NotNull Player player) {
+        return instance.levelSystemProvider.getSystem(player);
     }
 
-    public static PlayerUtils.ClassSystem getClassSystem() {
-        return instance.classSystem;
+    public static BukkitClassSystem getClassSystem(@NotNull Player player) {
+        return instance.classSystemProvider.getSystem(player);
     }
 
     @Nullable
@@ -125,7 +130,20 @@ public class RPGInventory extends PluginLifecycle {
     }
 
     @Override
+    public void onLoad() {
+        if (!checkMimicEnabled()) {
+            getServer()
+                    .getServicesManager()
+                    .register(BukkitItemsRegistry.class, new RPGInventoryItemsRegistry(), this, ServicePriority.High);
+        }
+    }
+
+    @Override
     public void onEnable() {
+        if (!initMimicSystems()) {
+            return;
+        }
+
         this.updateConfig();
         Config.reload();
         language = new FileLanguage(this);
@@ -182,14 +200,26 @@ public class RPGInventory extends PluginLifecycle {
                 .setExecutor(new TrackedCommandExecutor(new RPGInventoryCommandExecutor(), getReporter()));
 
         this.checkUpdates(null);
+    }
 
-        // Do this after all plugins loaded
-        new TrackedBukkitRunnable() {
-            @Override
-            public void run() {
-                checkThatSystemsLoaded();
-            }
-        }.runTask(this);
+    private boolean initMimicSystems() {
+        boolean isMimicEnabled = checkMimicEnabled();
+        if (isMimicEnabled) {
+            ServicesManager servicesManager = getServer().getServicesManager();
+            this.levelSystemProvider = servicesManager.load(BukkitLevelSystem.Provider.class);
+            Log.i("Level system ''{0}'' found.", this.levelSystemProvider.getId());
+            this.classSystemProvider = servicesManager.load(BukkitClassSystem.Provider.class);
+            Log.i("Class system ''{0}'' found.", this.classSystemProvider.getId());
+        } else {
+            Log.s("Mimic is required for RPGInventory to use levels and classes from other RPG plugins!");
+            Log.s("Download it from SpigotMC: https://www.spigotmc.org/resources/82515/");
+            getServer().getPluginManager().disablePlugin(this);
+        }
+        return isMimicEnabled;
+    }
+
+    private boolean checkMimicEnabled() {
+        return getServer().getPluginManager().isPluginEnabled("Mimic");
     }
 
     private boolean checkRequirements() {
@@ -220,35 +250,7 @@ public class RPGInventory extends PluginLifecycle {
             Log.w("Economy not found!");
         }
 
-        initLevelSystem();
-        initClassSystem();
-
         return InventoryManager.init(this) && SlotManager.init();
-    }
-
-    private void initLevelSystem() {
-        this.levelSystem = SafeEnums.valueOfOrDefault(PlayerUtils.LevelSystem.class,
-                Config.getConfig().getString("level-system"), PlayerUtils.LevelSystem.EXP, "level system");
-    }
-
-    private void initClassSystem() {
-        this.classSystem = SafeEnums.valueOfOrDefault(PlayerUtils.ClassSystem.class,
-                Config.getConfig().getString("class-system"), PlayerUtils.ClassSystem.PERMISSIONS, "class-system");
-    }
-
-    private void checkThatSystemsLoaded() {
-        PluginManager pm = this.getServer().getPluginManager();
-        if (levelSystem != PlayerUtils.LevelSystem.EXP && !pm.isPluginEnabled(levelSystem.getPluginName())) {
-            Log.w("Level-system {0} is not enabled!", levelSystem.getPluginName());
-            Log.w("Will be used EXP by default");
-            levelSystem = PlayerUtils.LevelSystem.EXP;
-        }
-
-        if (classSystem != PlayerUtils.ClassSystem.PERMISSIONS && !pm.isPluginEnabled(classSystem.getPluginName())) {
-            Log.w("Class-system {0} is not enabled!", classSystem.getPluginName());
-            Log.w("Will be used PERMISSIONS by default");
-            classSystem = PlayerUtils.ClassSystem.PERMISSIONS;
-        }
     }
 
     @Override